+++ contracts/loader/libraries/FeeERC20.sol
@@ -7,10 +7,10 @@ import "../../common/libraries/Fee.sol";
 library FeeERC20 {
 	using SafeERC20 for IERC20;
 
-	function transferGetFee(IERC20 token, address to, uint amount) internal returns (uint) {
+	function transferGetFee(IERC20 token, address to, uint amount) internal returns (uint16) {
 		uint preBalance = token.balanceOf(to);
 		token.safeTransfer(to, amount);
 		uint postBalance = token.balanceOf(to);
-		return Fee.get(amount, postBalance - preBalance);
+		return uint16(Fee.get(amount, postBalance - preBalance));
 	}
 }
